requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
mysql-connector-python==8.0.25
pandas==1.3.0
scikit-learn==0.24.2
schedule==1.1.0
numpy==1.21.0
